refactor: 버튼 템플릿 수정
