--- src/js/templates/saveButton.js
@@ -1,7 +1,0 @@
-export const saveButtonTemplate = () => {
-    return `<button class="btn save-button">⬇️ 저장</button>`;
-};
-
-export const saveCancelButtonTemplate = () => {
-    return `<button class="btn save-cancel-button">↪️ 저장 취소</button>`;
-};
